Indirect PID merge

+++ src/pidtask.h
@@ -3,8 +3,8 @@
  *
  * Purpose: Contain all of the structures and defines for implementing the
  *          PID algorythm in the PID task that will be launched in main.c
- *
- *
+ *          
+ *          
  *
  * Course:  ECE 544 - Embedded Systems Design, Winter 2025
  * Authors: Nikolay Nikolov, Ashten Bontrager
@@ -15,12 +15,12 @@
 
 /* PID structure definition */
 typedef struct {
-    uint32_t Kp;           // proportional gain
+    float Kp;           // proportional gain
     float Ki;           // integral gain
     float Kd;           // derivative gain
     uint16_t setpoint;  // intensity setpoint
     float integral;     // error acccumulator, sum of errors over time
-    uint16_t prev_error;   // previous error, used for (d_error/dt)
+    float prev_error;   // previous error, used for (d_error/dt)
     float delta_t;      // change in time for derivative
     uint16_t max_lim;      // value to set upper rail for output
     uint8_t min_lim;      // value to set lower rail for output
@@ -28,7 +28,7 @@ typedef struct {
 
 // prototype for function that initializes PID structure
 // returns byte used for checking if init has happened or not
-bool pid_init ( PID_t* pid );
+bool pid_init (PID_t* pid);
 
 // prototype of pid function that returns the controlled signal as a float
 // takes lux_value, returned from TSL2561 sensor returns float vlaue 
@@ -36,18 +36,18 @@ bool pid_init ( PID_t* pid );
 float pid_funct (PID_t* pid, uint16_t lux_value, uint8_t switches);
 
 /*********************PID Task Prototype*************************************
- *   Task Handles the Following:
- *   Reads perameter message from MsgQ
- *   Update new control/setpoint parameters
- *   Get Current lux readig from TSL2561 sensor
- *       done using the TSL2561 driver in implemented
- *       in the C file of the same name
- *   Execute PID algo function
- *   Drive PWM signal for LED, use RGB writ commands
- *   write to display thread MsgQ to update
- *   setpoint and current lux
- *****************************************************************************/
-void PID_Task ( void* p );
+*   Task Handles the Following:
+*   Reads perameter message from MsgQ
+*   Update new control/setpoint parameters
+*   Get Current lux readig from TSL2561 sensor
+*       done using the TSL2561 driver in implemented
+*       in the C file of the same name
+*   Execute PID algo function 
+*   Drive PWM signal for LED, use RGB writ commands
+*   write to display thread MsgQ to update
+*   setpoint and current lux
+*****************************************************************************/
+void PID_Task (void* p);
 
 /************************Display Task****************************************
 *   Gets lux and setpoint values from Q and updates 7-seg display
